thermal flicker
