+++ Dashboard.desktop
@@ -1,0 +1,6 @@
+[Desktop Entry]
+Exec=sudo python3 /home/sensor/Desktop/coviduipy/main.py
+Type=Application
+Terminal=true
+Icon=/home/sensor/Desktop/coviduipy/coronavirus.png
+Name[en_US]=Dashboard
