+++ uvc-deviceinfo.py
@@ -1,0 +1,37 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from uvctypes import *
+
+def main():
+  ctx = POINTER(uvc_context)()
+  dev = POINTER(uvc_device)()
+  devh = POINTER(uvc_device_handle)()
+  ctrl = uvc_stream_ctrl()
+
+  res = libuvc.uvc_init(byref(ctx), 0)
+  if res < 0:
+    print("uvc_init error")
+    exit(1)
+
+  try:
+    res = libuvc.uvc_find_device(ctx, byref(dev), PT_USB_VID, PT_USB_PID, 0)
+    if res < 0:
+      print("uvc_find_device error")
+      exit(1)
+
+    try:
+      res = libuvc.uvc_open(dev, byref(devh))
+      if res < 0:
+        print("uvc_open error")
+        exit(1)
+
+      print_device_info(devh)
+
+    finally:
+      libuvc.uvc_unref_device(dev)
+  finally:
+    libuvc.uvc_exit(ctx)
+
+if __name__ == '__main__':
+  main()
